renamed addJobLead to addJobPosting
